Gerando gráfico inicial

+++ algorithms.h
@@ -1,6 +1,6 @@
-void insertion_sort(long int*, long int);
-void bubble_sort(long int*, long int);
-void selection_sort(long int*, long int);
-void merge_sort(long int*, long int);
-void quick_sort(long int*, long int);
-void heap_sort(long int*, long int);
+void insertion_sort(int *, int);
+void bubble_sort(int *, int);
+void selection_sort(int *, int);
+void merge_sort(int *, int);
+void quick_sort(int *, int);
+void heap_sort(int *, int);
